requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycodestyle==2.5.0
pymongo==3.10.0
python-dateutil==2.8.1
six==1.13.0
Werkzeug==0.16.0
